Factor out option compatibility checks to JellyOptions
This will be very useful for gRPC service implementations. The example of using Jelly gRPC was updated to reflect this.
